Extracts canvas parameters out to own class
