+++ lib/scripts/arc.js
@@ -11,7 +11,7 @@ let Arc = function(tileNumber, orientation) {
   this.shapeType = "arc";
 };
 
-const cardinalDirections = ["left", "right", "top", "bottom"];
+let cardinalDirections = ["left", "right", "top", "bottom"];
 
 let startingOrientation = function(winningOrientation) {
   let result = cardinalDirections.sample();
