+++ lib/scripts/doublearc.js
@@ -5,7 +5,7 @@ let doubleArc = function(tileNumber, orientation) {
   this.shapeType = "doubleArc";
 };
 
-const cardinalDirections = ["top", "right", "bottom", "left"];
+let cardinalDirections = ["top", "right", "bottom", "left"];
 
 let startingOrientation = function(winningOrientation) {
   let result = cardinalDirections.sample();
