+++ lib/scripts/effects.js
@@ -1,4 +1,5 @@
 const $ = require('jquery');
+
 let Effects = function() {};
 
 let canvasColors = ['#59323C', '#320132', '#8E3F25', '#443B22', '#1E280B', '#4D3933'];
@@ -37,7 +38,7 @@ Effects.prototype.introSequence = function(self) {
     $('.level-announcements').fadeOut(1000);
     $('canvas').delay(2000).fadeIn(1000);
     $('.game-div').off('click');
-    self.playLevel(0);
+    self.playLevel(9);
   });
 };
