+++ lib/scripts/game.js
@@ -14,17 +14,12 @@ let Game = function(context, canvas) {
   this.context = context;
   this.canvas = canvas;
   this.width = canvas.width;
-  this.height = canvas.height;
   this.context.lineWidth = 4;
   this.objectIndex = {};
-  let renderer = new Renderer(this);
-  this.renderer = renderer;
-  let grid = new Grid();
-  this.grid = grid;
-  let solutions = new Solutions(this);
-  this.solutions = solutions;
-  let effects = new Effects();
-  this.effects = effects;
+  this.renderer = new Renderer(this);
+  this.grid = new Grid();
+  this.solutions = new Solutions(this);
+  this.effects = new Effects();
   this.counter = 0;
 };
 
@@ -156,7 +151,7 @@ let startNextLevel = function(self, levelNumber) {
   self.playLevel(levelNumber + 1);
 };
 
-const rotations = {
+let rotations = {
   "left": "top",
   "top": "right",
   "right": "bottom",
