+++ lib/scripts/line.js
@@ -5,7 +5,7 @@ let Line = function(tileNumber, orientation) {
   this.shapeType = "line";
 };
 
-const cardinalDirections = ["horizontal", "vertical"];
+let cardinalDirections = ["horizontal", "vertical"];
 
 let startingOrientation = function(winningOrientation) {
   let result = cardinalDirections.sample();
