+++ lib/scripts/pattern-generator.js
@@ -1,4 +1,5 @@
 const Tadpole = require('./tadpole');
+const CanvasParameters = require('./canvas-parameters');
 
 let PatternGenerator = function(game) {
   this.game = game;
@@ -34,120 +35,13 @@ let createShapeDetails = function(tadpoleIndex) {
   return arrayOfTileDetails;
 };
 
-let smallCanvas = {
-  "E4": ["n", "u", "u", "n"],
-  "F4": ["n", "u", "u", "u"],
-  "G4": ["n", "u", "u", "u"],
-  "H4": ["n", "n", "u", "u"],
-  "E5": ["u", "u", "u", "n"],
-  "F5": ["u", "u", "u", "u"],
-  "G5": ["u", "u", "u", "u"],
-  "H5": ["u", "n", "u", "u"],
-  "E6": ["u", "u", "n", "n"],
-  "F6": ["u", "u", "n", "u"],
-  "G6": ["u", "u", "n", "u"],
-  "H6": ["u", "n", "n", "u"]
-};
-
-let mediumCanvas = {
-  "D3": ["n", "u", "u", "n"],
-  "D4": ["u", "u", "u", "n"],
-  "D5": ["u", "u", "u", "n"],
-  "D6": ["u", "u", "u", "n"],
-  "D7": ["u", "u", "n", "n"],
-  "E3": ["n", "u", "u", "u"],
-  "E4": ["u", "u", "u", "u"],
-  "E5": ["u", "u", "u", "u"],
-  "E6": ["u", "u", "u", "u"],
-  "E7": ["u", "u", "n", "u"],
-  "F3": ["n", "u", "u", "u"],
-  "F4": ["u", "u", "u", "u"],
-  "F5": ["u", "u", "u", "u"],
-  "F6": ["u", "u", "u", "u"],
-  "F7": ["u", "u", "n", "u"],
-  "G3": ["n", "u", "u", "u"],
-  "G4": ["u", "u", "u", "u"],
-  "G5": ["u", "u", "u", "u"],
-  "G6": ["u", "u", "u", "u"],
-  "G7": ["u", "u", "n", "u"],
-  "H3": ["n", "u", "u", "u"],
-  "H4": ["u", "u", "u", "u"],
-  "H5": ["u", "u", "u", "u"],
-  "H6": ["u", "u", "u", "u"],
-  "H7": ["u", "u", "n", "u"],
-  "I3": ["n", "n", "u", "u"],
-  "I4": ["u", "n", "u", "u"],
-  "I5": ["u", "n", "u", "u"],
-  "I6": ["u", "n", "u", "u"],
-  "I7": ["u", "n", "n", "u"]
-};
-
-let largeCanvas = {
-  "C2": ["n", "u", "u", "n"],
-  "C3": ["u", "u", "u", "n"],
-  "C4": ["u", "u", "u", "n"],
-  "C5": ["u", "u", "u", "n"],
-  "C6": ["u", "u", "u", "n"],
-  "C7": ["u", "u", "u", "n"],
-  "C8": ["u", "u", "n", "n"],
-  "D2": ["n", "u", "u", "u"],
-  "D3": ["u", "u", "u", "u"],
-  "D4": ["u", "u", "u", "u"],
-  "D5": ["u", "u", "u", "u"],
-  "D6": ["u", "u", "u", "u"],
-  "D7": ["u", "u", "u", "u"],
-  "D8": ["u", "u", "n", "u"],
-  "E2": ["n", "u", "u", "u"],
-  "E3": ["u", "u", "u", "u"],
-  "E4": ["u", "u", "u", "u"],
-  "E5": ["u", "u", "u", "u"],
-  "E6": ["u", "u", "u", "u"],
-  "E7": ["u", "u", "u", "u"],
-  "E8": ["u", "u", "n", "u"],
-  "F2": ["n", "u", "u", "u"],
-  "F3": ["u", "u", "u", "u"],
-  "F4": ["u", "u", "u", "u"],
-  "F5": ["u", "u", "u", "u"],
-  "F6": ["u", "u", "u", "u"],
-  "F7": ["u", "u", "u", "u"],
-  "F8": ["u", "u", "n", "u"],
-  "G2": ["n", "u", "u", "u"],
-  "G3": ["u", "u", "u", "u"],
-  "G4": ["u", "u", "u", "u"],
-  "G5": ["u", "u", "u", "u"],
-  "G6": ["u", "u", "u", "u"],
-  "G7": ["u", "u", "u", "u"],
-  "G8": ["u", "u", "n", "u"],
-  "H2": ["n", "u", "u", "u"],
-  "H3": ["u", "u", "u", "u"],
-  "H4": ["u", "u", "u", "u"],
-  "H5": ["u", "u", "u", "u"],
-  "H6": ["u", "u", "u", "u"],
-  "H7": ["u", "u", "u", "u"],
-  "H8": ["u", "u", "n", "u"],
-  "I2": ["n", "u", "u", "u"],
-  "I3": ["u", "u", "u", "u"],
-  "I4": ["u", "u", "u", "u"],
-  "I5": ["u", "u", "u", "u"],
-  "I6": ["u", "u", "u", "u"],
-  "I7": ["u", "u", "u", "u"],
-  "I8": ["u", "u", "n", "u"],
-  "J2": ["n", "n", "u", "u"],
-  "J3": ["u", "n", "u", "u"],
-  "J4": ["u", "n", "u", "u"],
-  "J5": ["u", "n", "u", "u"],
-  "J6": ["u", "n", "u", "u"],
-  "J7": ["u", "n", "u", "u"],
-  "J8": ["u", "n", "n", "u"],
-};
-
 PatternGenerator.prototype.canvasMatcher = function(levelNumber) {
   if(levelNumber <= 10) {
-    return smallCanvas;
+    return new CanvasParameters("small");
   } else if(levelNumber > 10 && levelNumber <= 15) {
-    return mediumCanvas;
+    return new CanvasParameters("medium");
   } else {
-    return largeCanvas;
+    return new CanvasParameters("large");
   }
 };
