+++ lib/scripts/renderer.js
@@ -6,8 +6,7 @@ let Renderer = function(game) {
   this.spokeLineCoordinates = spokeLineCoordinates;
   this.lineCoordinates = lineCoordinates;
   this.doubleArcCoordinates = doubleArcCoordinates;
-  let grid = new Grid();
-  this.grid = grid;
+  this.grid = new Grid();
 };
 
 Renderer.prototype.renderTile = function(shape, location, orientation) {
@@ -41,7 +40,7 @@ Renderer.prototype.renderSpoke = function(location, orientation) {
   this.context.clearRect(startX, startY, 40, 40);
   this.context.strokeStyle = "#BFAF80";
   this.context.beginPath();
-  this.context.arc(startX + 20, startY + 20, 10, Math.PI * 0.0, Math.PI * 2.0, false);
+  this.context.arc(startX + 20, startY + 20, 10, 0, (Math.PI * 2.0), false);
   this.context.stroke();
 
   this.context.strokeStyle = "#BFAF80";
@@ -55,14 +54,14 @@ Renderer.prototype.renderArc = function(location, orientation) {
   let startX = this.findStartX(location);
   let startY = this.findStartY(location);
 
-  const radians = {
+  let radians = {
     "top": [0.0, 1.5],
     "right": [0.5, 0.0],
     "bottom": [1.0, 0.5],
     "left": [1.5, 1.0]
   };
 
-  const startPoints = {
+  let startPoints = {
     "top": [0, 40],
     "right": [0, 0],
     "bottom": [40, 0],
@@ -110,34 +109,34 @@ Renderer.prototype.renderQuadArc = function(location) {
   this.context.clearRect(startX, startY, 40, 40);
   this.context.strokeStyle = "#BFAF80";
   this.context.beginPath();
-  this.context.arc(startX + 0, startY + 40, 20, (Math.PI * 1.5), (Math.PI * 0.0), false);
+  this.context.arc(startX, startY, 20, (Math.PI * 1.5), Math.PI, false);
+  this.context.stroke();
+  this.context.beginPath();
+  this.context.arc(startX + 40, startY, 20, (Math.PI * 0.5), Math.PI, false);
   this.context.stroke();
   this.context.beginPath();
   this.context.arc(startX + 40, startY + 40, 20, Math.PI, (Math.PI * 1.5), false);
   this.context.stroke();
   this.context.beginPath();
-  this.context.arc(startX + 0, startY + 0, 20, (Math.PI * 1.5), (Math.PI * 1.0), false);
-  this.context.stroke();
-  this.context.beginPath();
-  this.context.arc(startX + 40, startY + 0, 20, (Math.PI * 0.5), (Math.PI * 1.0), false);
+  this.context.arc(startX, startY + 40, 20, (Math.PI * 1.5), 0, false);
   this.context.stroke();
 };
 
-const doubleArcCoordinates = {
-  "top": [[40, 0, 0.5, 1.0], [0, 0, 0.0, 0.5]],
+let doubleArcCoordinates = {
+  "top": [[40, 0, 0.5, 1.0], [0, 0, 0, 0.5]],
   "right": [[40, 40, 1.0, 1.5], [40, 0, 0.5, 1.0]],
-  "bottom": [[0, 40, 1.5, 0.0], [40, 40, 1.0, 1.5]],
-  "left": [[0, 0, 0.0, 0.5], [0, 40, 1.5, 0.0]]
+  "bottom": [[0, 40, 1.5, 0], [40, 40, 1.0, 1.5]],
+  "left": [[0, 0, 0, 0.5], [0, 40, 1.5, 0]]
 };
 
-const spokeLineCoordinates = {
+let spokeLineCoordinates = {
   "top": [[20,0], [20,10]],
   "right": [[30,20], [40,20]],
   "bottom": [[20,30], [20,40]],
   "left": [[0,20], [10,20]]
 };
 
-const lineCoordinates = {
+let lineCoordinates = {
   "vertical": [[20,0], [20,40]],
   "horizontal": [[0,20], [40,20]]
 };
