+++ lib/scripts/solutions.js
@@ -2,8 +2,7 @@ const PatternGenerator = require('./pattern-generator');
 
 let Solutions = function(game) {
   this.game = game;
-  let patternGenerator = new PatternGenerator();
-  this.patternGenerator = patternGenerator;
+  this.patternGenerator = new PatternGenerator();
 };
 
 Solutions.prototype.setArrangement = function(levelNumber) {
