+++ lib/scripts/spoke.js
@@ -5,7 +5,7 @@ let Spoke = function(tileNumber, orientation) {
   this.shapeType = "spoke";
 };
 
-const cardinalDirections = ["left", "right", "top", "bottom"];
+let cardinalDirections = ["left", "right", "top", "bottom"];
 
 let startingOrientation = function(winningOrientation) {
   let result = cardinalDirections.sample();
