+++ lib/scripts/tadpole.js
@@ -133,16 +133,37 @@ let meetsAllConditions = function(allOptionsWithDuplicates) {
   return matches;
 };
 
-
 let allOptions = [
-  ["spoke", "top"],["spoke", "right"],["spoke", "bottom"],["spoke", "left"],["arc", "top"],["arc", "right"],["arc", "bottom"],["arc", "left"],["line", "horizontal"],["line", "vertical"],["doubleArc", "top"],["doubleArc", "right"],["doubleArc", "bottom"],["doubleArc", "left"],["quadArc", "all"]
+  ["spoke", "top"],
+  ["spoke", "right"],
+  ["spoke", "bottom"],
+  ["spoke", "left"],
+  ["arc", "top"],
+  ["arc", "right"],
+  ["arc", "bottom"],
+  ["arc", "left"],
+  ["line", "horizontal"],
+  ["line", "vertical"],
+  ["doubleArc", "top"],
+  ["doubleArc", "right"],
+  ["doubleArc", "bottom"],
+  ["doubleArc", "left"],
+  ["quadArc", "all"]
 ];
 
 let tileOptions = {
-  "top": { "y": [["spoke", "top"],["arc", "right"],["arc", "bottom"],["line", "vertical"],["doubleArc", "top"],["doubleArc", "right"],["doubleArc", "left"],["quadArc", "all"]], "n": [["spoke", "right"],["spoke", "bottom"],["spoke", "left"],["arc", "top"],["arc", "left"],["line", "horizontal"],["doubleArc", "bottom"]], "u": allOptions },
-  "right": { "y": [["spoke", "right"],["arc", "bottom"],["arc", "left"],["line", "horizontal"],["doubleArc", "top"],["doubleArc", "right"],["doubleArc", "bottom"],["quadArc", "all"]], "n": [["spoke", "top"],["spoke", "bottom"],["spoke", "left"],["arc", "top"],["arc", "right"],["line", "vertical"],["doubleArc", "left"]], "u": allOptions },
-  "bottom": { "y": [["spoke", "bottom"],["arc", "top"],["arc", "left"],["line", "vertical"],["doubleArc", "right"],["doubleArc", "bottom"],["doubleArc", "left"],["quadArc", "all"]], "n": [["spoke", "top"],["spoke", "right"],["spoke", "left"],["arc", "right"],["arc", "bottom"],["line", "horizontal"],["doubleArc", "top"]], "u": allOptions },
-  "left": { "y": [["spoke", "left"],["arc", "top"],["arc", "right"],["line", "horizontal"],["doubleArc", "top"],["doubleArc", "bottom"],["doubleArc", "left"],["quadArc", "all"]], "n": [["spoke", "top"],["spoke", "right"],["spoke", "bottom"],["arc", "bottom"],["arc", "left"],["line", "vertical"],["doubleArc", "right"]], "u": allOptions }
+  "top":    { "y": [["spoke", "top"],["arc", "right"],["arc", "bottom"],["line", "vertical"],["doubleArc", "top"],["doubleArc", "right"],["doubleArc", "left"],["quadArc", "all"]],
+              "n": [["spoke", "right"],["spoke", "bottom"],["spoke", "left"],["arc", "top"],["arc", "left"],["line", "horizontal"],["doubleArc", "bottom"]],
+              "u": allOptions },
+  "right":  { "y": [["spoke", "right"],["arc", "bottom"],["arc", "left"],["line", "horizontal"],["doubleArc", "top"],["doubleArc", "right"],["doubleArc", "bottom"],["quadArc", "all"]], 
+              "n": [["spoke", "top"],["spoke", "bottom"],["spoke", "left"],["arc", "top"],["arc", "right"],["line", "vertical"],["doubleArc", "left"]],
+              "u": allOptions },
+  "bottom": { "y": [["spoke", "bottom"],["arc", "top"],["arc", "left"],["line", "vertical"],["doubleArc", "right"],["doubleArc", "bottom"],["doubleArc", "left"],["quadArc", "all"]],
+              "n": [["spoke", "top"],["spoke", "right"],["spoke", "left"],["arc", "right"],["arc", "bottom"],["line", "horizontal"],["doubleArc", "top"]],
+              "u": allOptions },
+  "left":   { "y": [["spoke", "left"],["arc", "top"],["arc", "right"],["line", "horizontal"],["doubleArc", "top"],["doubleArc", "bottom"],["doubleArc", "left"],["quadArc", "all"]],
+              "n": [["spoke", "top"],["spoke", "right"],["spoke", "bottom"],["arc", "bottom"],["arc", "left"],["line", "vertical"],["doubleArc", "right"]],
+              "u": allOptions }
 };
 
 module.exports = Tadpole;
